+++ Web/Controllers/QuestionsController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -93,6 +94,11 @@ namespace Web.Controllers
 
                 return View(questionViewModel);
             }
+            catch(DbUpdateException dbex)
+            {
+                ViewData["ErrorMessage"] = "DB issue - " + dbex.Message;
+                return View("Error");
+            }
             catch(Exception ex)
             {
                 ViewData["ErrorMessage"] = ex.Message;
@@ -104,7 +110,7 @@ namespace Web.Controllers
         [HttpGet]
         public IActionResult AddQuestion()
         {
-            return View();
+            return View();            
         }
 
         // POST: QuestionsController/AddQuestion
@@ -114,22 +120,36 @@ namespace Web.Controllers
         {
             if (ModelState.IsValid)
             {
-                var currentlySignedInUser = await _userManager.GetUserAsync(User);
-                questionViewModel.UserId = currentlySignedInUser.Id;
-                string uniqueFileName = null;
-                if (questionViewModel.Image != null && _fileOperations.ValidateImageType(questionViewModel.Image.FileName) == true)
+                try
                 {
-                    // for more advanced projects add a composite file provider - for now wwwroot
-                    // https://docs.microsoft.com/en-us/aspnet/core/fundamentals/file-providers?view=aspnetcore-5.0#compositefileprovider
-                    string serverImagesDirectory = Path.Combine(_webHostEnvironment.WebRootPath, "uploads");
-                    uniqueFileName = _fileOperations.AssembleQuestionUploadedFileName(questionViewModel.UserId, questionViewModel.Image.FileName);
-                    string filePath = Path.Combine(serverImagesDirectory, uniqueFileName);
-                    await questionViewModel.Image.CopyToAsync(new FileStream(filePath, FileMode.Create));
+                    var currentlySignedInUser = await _userManager.GetUserAsync(User);
+                    questionViewModel.UserId = currentlySignedInUser.Id;
+                    string uniqueFileName = null;
+                    if (questionViewModel.Image != null && _fileOperations.ValidateImageType(questionViewModel.Image.FileName) == true)
+                    {
+                        // for more advanced projects add a composite file provider - for now wwwroot
+                        // https://docs.microsoft.com/en-us/aspnet/core/fundamentals/file-providers?view=aspnetcore-5.0#compositefileprovider
+                        string serverImagesDirectory = Path.Combine(_webHostEnvironment.WebRootPath, "uploads");
+                        uniqueFileName = _fileOperations.AssembleQuestionUploadedFileName(questionViewModel.UserId, questionViewModel.Image.FileName);
+                        string filePath = Path.Combine(serverImagesDirectory, uniqueFileName);
+                        await questionViewModel.Image.CopyToAsync(new FileStream(filePath, FileMode.Create));
+                    }
+                    var question = _mapper.Map<QuestionViewModel, Question>(questionViewModel);
+                    question.ImageNamePath = uniqueFileName;
+                    var resultQuestion = await _repository.AddQuestionAsync(question);
+                    return RedirectToAction("Details", new { questionId = resultQuestion.Id });
                 }
-                var question = _mapper.Map<QuestionViewModel, Question>(questionViewModel);
-                question.ImageNamePath = uniqueFileName;
-                var resultQuestion = await _repository.AddQuestionAsync(question);
-                return RedirectToAction("Details", new { questionId = resultQuestion.Id });
+                catch (DbUpdateException dbex)
+                {
+                    ViewData["ErrorMessage"] = "DB issue - " + dbex.Message;
+                    return View("Error");
+                }
+                catch (Exception ex)
+                {
+                    ViewData["ErrorMessage"] = ex.Message;
+                    return View("Error");
+                }
+                
             }
             return View();
         }
@@ -137,14 +157,21 @@ namespace Web.Controllers
         // GET: QuestionsController/5/Edit
         [HttpGet]
         [Route("questions/{questionId}/edit")]
-        public async Task<IActionResult> Edit(int questionId)
+        public async Task<IActionResult> EditQuestion(int questionId)
         {
             var question = await _repository.GetQuestionByIdWithoutDetailsAsync(questionId);
+            if (question == null)
+            {
+                Response.StatusCode = 404;
+                ViewData["ErrorMessage"] = "404 Resource not found.";
+                return View("Error");
+            }
             if (String.Equals(User.FindFirstValue(ClaimTypes.NameIdentifier), question.UserId) == false)
             {
                 return RedirectToAction("AccessDenied", "Account");
             }
             var questionViewModel = _mapper.Map<Question, QuestionViewModel>(question);
+            //ViewData["ValidationErrorMessage"] = validationErrorMessage;
             return View(questionViewModel);
         }
 
@@ -152,47 +179,89 @@ namespace Web.Controllers
         [HttpPost]
         [Route("questions/{questionId}/edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(QuestionViewModel questionViewModel)
+        public async Task<IActionResult> EditQuestion(QuestionViewModel questionViewModel)
         {
             var question = await _repository.GetQuestionByIdWithoutDetailsAsync(questionViewModel.Id);
+            if (question == null)
+            {
+                Response.StatusCode = 404;
+                ViewData["ErrorMessage"] = "404 Resource not found.";
+                return View("Error");
+            }
             if (String.Equals(User.FindFirstValue(ClaimTypes.NameIdentifier), question.UserId) == false)
             {
                 return RedirectToAction("AccessDenied", "Account");
             }
             if (ModelState.IsValid)
             {
-                var currentlySignedInUser = await _userManager.GetUserAsync(User);
-                questionViewModel.UserId = currentlySignedInUser.Id;
-                string uniqueFileName = null;
-                if (questionViewModel.Image != null)
+                try
                 {
-                    // for more advanced projects add a composite file provider - for now wwwroot
-                    // https://docs.microsoft.com/en-us/aspnet/core/fundamentals/file-providers?view=aspnetcore-5.0#compositefileprovider
-                    string serverImagesDirectory = Path.Combine(_webHostEnvironment.WebRootPath, "uploads");
-                    uniqueFileName = _fileOperations.AssembleQuestionUploadedFileName(questionViewModel.UserId, questionViewModel.Image.FileName);
-                    string filePath = Path.Combine(serverImagesDirectory, uniqueFileName);
-                    await questionViewModel.Image.CopyToAsync(new FileStream(filePath, FileMode.Create));
+                    var currentlySignedInUser = await _userManager.GetUserAsync(User);
+                    questionViewModel.UserId = currentlySignedInUser.Id;
+                    string uniqueFileName = null;
+                    if (questionViewModel.Image != null)
+                    {
+                        // for more advanced projects add a composite file provider - for now wwwroot
+                        // https://docs.microsoft.com/en-us/aspnet/core/fundamentals/file-providers?view=aspnetcore-5.0#compositefileprovider
+                        string serverImagesDirectory = Path.Combine(_webHostEnvironment.WebRootPath, "uploads");
+                        uniqueFileName = _fileOperations.AssembleQuestionUploadedFileName(questionViewModel.UserId, questionViewModel.Image.FileName);
+                        string filePath = Path.Combine(serverImagesDirectory, uniqueFileName);
+                        await questionViewModel.Image.CopyToAsync(new FileStream(filePath, FileMode.Create));
+                    }
+                    question = _mapper.Map<QuestionViewModel, Question>(questionViewModel);
+                    question.ImageNamePath = uniqueFileName;
+                    await _repository.EditQuestionAsync(question);
+                    return RedirectToAction("Details", new { questionId = questionViewModel.Id });
                 }
-                question = _mapper.Map<QuestionViewModel, Question>(questionViewModel);
-                question.ImageNamePath = uniqueFileName;
-                await _repository.EditQuestionAsync(question);
-                return RedirectToAction("Details", new { questionId = questionViewModel.Id });
+                catch (DbUpdateException dbex)
+                {
+                    ViewData["ErrorMessage"] = "DB issue - " + dbex.Message;
+                    return View("Error");
+                }
+                catch (Exception ex)
+                {
+                    ViewData["ErrorMessage"] = ex.Message;
+                    return View("Error");
+                }
             }
-            return View(questionViewModel.Id);
+            // simply returning View somehow sends a Post request and parameter type error shows up as it expects a QVM
+            // did some research, stackoverflow old 6 and 3 yr questions - no answers
+            // also set up a custom error message via ViewData but this is more "expected behavior" as it sneds a Get, probably
+            // answers controller has no issues with this
+            return await EditQuestion(questionViewModel.Id);
         }
 
         // Get: QuestionsController/5/Remove
         [HttpGet]
         [Route("questions/remove/{questionId}")]
-        public async Task<IActionResult> Remove(int questionId)
+        public async Task<IActionResult> RemoveQuestion(int questionId)
         {
             var question = await _repository.GetQuestionByIdWithoutDetailsAsync(questionId);
+            if (question == null)
+            {
+                Response.StatusCode = 404;
+                ViewData["ErrorMessage"] = "404 Resource not found.";
+                return View("Error");
+            }
             if (String.Equals(User.FindFirstValue(ClaimTypes.NameIdentifier), question.UserId) == false)
             {
                 return RedirectToAction("AccessDenied", "Account");
             }
-            await _repository.RemoveQuestionById(questionId);
-            return RedirectToAction("Index", "List");
+            try
+            {
+                await _repository.RemoveQuestionById(questionId);
+                return RedirectToAction("Index", "List");
+            }
+            catch (DbUpdateException dbex)
+            {
+                ViewData["ErrorMessage"] = "DB issue - " + dbex.Message;
+                return View("Error");
+            }
+            catch (Exception ex)
+            {
+                ViewData["ErrorMessage"] = ex.Message;
+                return View("Error");
+            }
         }
 
         // Get: QuestionsController/5/RemoveImage
@@ -201,45 +270,112 @@ namespace Web.Controllers
         public async Task<IActionResult> RemoveImage(int questionId)
         {
             var question = await _repository.GetQuestionByIdWithoutDetailsAsync(questionId);
+            if (question == null)
+            {
+                Response.StatusCode = 404;
+                ViewData["ErrorMessage"] = "404 Resource not found.";
+                return View("Error");
+            }
             if (String.Equals(User.FindFirstValue(ClaimTypes.NameIdentifier), question.UserId) == false)
             {
                 return RedirectToAction("AccessDenied", "Account");
             }
-            await _repository.RemoveQuestionImageByQuestionId(questionId);
-            return RedirectToAction("Details", new { questionId = questionId });
+            try
+            {
+                await _repository.RemoveQuestionImageByQuestionId(questionId);
+                return RedirectToAction("Details", new { questionId = questionId });
+            }
+            catch (DbUpdateException dbex)
+            {
+                ViewData["ErrorMessage"] = "DB issue - " + dbex.Message;
+                return View("Error");
+            }
+            catch (Exception ex)
+            {
+                ViewData["ErrorMessage"] = ex.Message;
+                return View("Error");
+            }
         }
 
         // Get: QuestionsController/5/VoteUp
         [HttpGet]
         [Route("questions/{questionId}/voteup")]
-        public async Task<IActionResult> VoteUpQuestion(int questionId, string redirection="list")
+        public async Task<IActionResult> VoteUpQuestion(int questionId, string redirection= "redirectToAllQuestions")
         {
-            await _repository.VoteUpQuestionById(questionId);
-            if (String.Equals("redirectToDetails", redirection))
+            var question = await _repository.GetQuestionByIdWithoutDetailsAsync(questionId);
+            if (question == null)
             {
-                return RedirectToAction("Details", new { questionId = questionId });
+                Response.StatusCode = 404;
+                ViewData["ErrorMessage"] = "404 Resource not found.";
+                return View("Error");
             }
-            else
+            try
             {
-                return RedirectToAction("Index", "List");
+                await _repository.VoteUpQuestionById(questionId);
+                if (String.Equals("redirectToDetails", redirection))
+                {
+                    return RedirectToAction("Details", new { questionId = questionId });
+                }
+                else if (String.Equals("redirectToHome", redirection))
+                {
+                    return RedirectToAction("Index", "Home");
+                }
+                else
+                {
+                    return RedirectToAction("AllQuestions", "Home");
+                }
             }
+            catch (DbUpdateException dbex)
+            {
+                ViewData["ErrorMessage"] = "DB issue - " + dbex.Message;
+                return View("Error");
+            }
+            catch (Exception ex)
+            {
+                ViewData["ErrorMessage"] = ex.Message;
+                return View("Error");
+            }
+            
         }
 
         // Get: QuestionsController/5/VoteDown
         [HttpGet]
         [Route("questions/{questionId}/votedown")]
-        public async Task<IActionResult> VoteDownQuestion(int questionId, string redirection="list")
+        public async Task<IActionResult> VoteDownQuestion(int questionId, string redirection="redirectToAllQuestions")
         {
-            await _repository.VoteDownQuestionById(questionId);
-            if (String.Equals("redirectToDetails", redirection))
+            var question = await _repository.GetQuestionByIdWithoutDetailsAsync(questionId);
+            if (question == null)
             {
-                return RedirectToAction("Details", new { questionId = questionId });
+                Response.StatusCode = 404;
+                ViewData["ErrorMessage"] = "404 Resource not found.";
+                return View("Error");
             }
-            else
+            try
             {
-                return RedirectToAction("Index", "List");
+                await _repository.VoteDownQuestionById(questionId);
+                if (String.Equals("redirectToDetails", redirection))
+                {
+                    return RedirectToAction("Details", new { questionId = questionId });
+                }
+                else if (String.Equals("redirectToHome", redirection))
+                {
+                    return RedirectToAction("Index", "Home");
+                }
+                else
+                {
+                    return RedirectToAction("AllQuestions", "Home");
+                }
             }
-            
+            catch (DbUpdateException dbex)
+            {
+                ViewData["ErrorMessage"] = "DB issue - " + dbex.Message;
+                return View("Error");
+            }
+            catch (Exception ex)
+            {
+                ViewData["ErrorMessage"] = ex.Message;
+                return View("Error");
+            }
         }
     }
 }
